% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy2.R
\name{dummy2}
\alias{dummy2}
\title{This is the title of dummy function}
\usage{
dummy2(arg1, arg2)
}
\arguments{
\item{arg1}{description of arg1}

\item{arg2}{description of arg2}
}
\value{
this function returns object of class \code{\link{R6Class}}.
}
\description{
This is the detailed description of this dummy function
}
\examples{
dummyOut = dummy("Shashank", "Good job.")
}
