% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello function}
\usage{
hello(x)
}
\arguments{
\item{x}{someName}
}
\value{
The output from \code{\link{print}}
}
\description{
This function prints "hello world".
}
\examples{
hello("Shashank")
\dontrun{
hello("Shashank")
}
}
