% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParaDRAM.R
\name{ParaDRAM}
\alias{ParaDRAM}
\title{Instantiate a \code{\link{R6}} \strong{\code{ParaDRAM}} class object for \strong{\link{paramonte}} package}
\value{
This function returns a \code{\link{R6}} \code{ParaDRAM} class object.
}
\description{
This is the \code{ParaDRAM} class to generate instances of \strong{serial} and \strong{parallel} \cr
\strong{Delayed-Rejection Adaptive Metropolis-Hastings Markov Chain Monte Carlo} sampler class \cr
of the \strong{\link{paramonte}} library. The \code{ParaDRAM} class is a child of the \cr
\code{ParaMonteSampler} class. \cr
The object of this class can be instantiated only via \code{\link{paramonte_class}} object \cr
method (for example: \code{pm$ParaDRAM()}).\cr
\cr
All \code{ParaDRAM} class attributes are optional and all attributes can be set after a \code{ParaDRAM} \cr
instance is returned by the class constructor.
}
\examples{

## ------------------------------------------------
## Method `ParaDRAM$new`
## ------------------------------------------------

    pm   = paramonte_class$new()
    pmpd = pm$ParaDRAM()

## ------------------------------------------------
## Method `ParaDRAM$print`
## ------------------------------------------------

    pm   = paramonte_class$new()
    pmpd = pm$ParaDRAM()
    pmpd$print()
    # or
    pm   = paramonte_class$new()
    pmpd = pm$ParaDRAM()
    pmpd
}
\section{Super class}{
\code{\link[paramonte:ParaMonteSampler]{paramonte::ParaMonteSampler}} -> \code{ParaDRAM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParaDRAM$new()}}
\item \href{#method-print}{\code{ParaDRAM$print()}}
\item \href{#method-clone}{\code{ParaDRAM$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paramonte" data-topic="ParaMonteSampler" data-id="readChain">}\href{../../paramonte/html/ParaMonteSampler.html#method-readChain}{\code{paramonte::ParaMonteSampler$readChain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paramonte" data-topic="ParaMonteSampler" data-id="readMarkovChain">}\href{../../paramonte/html/ParaMonteSampler.html#method-readMarkovChain}{\code{paramonte::ParaMonteSampler$readMarkovChain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paramonte" data-topic="ParaMonteSampler" data-id="readProgress">}\href{../../paramonte/html/ParaMonteSampler.html#method-readProgress}{\code{paramonte::ParaMonteSampler$readProgress()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paramonte" data-topic="ParaMonteSampler" data-id="readReport">}\href{../../paramonte/html/ParaMonteSampler.html#method-readReport}{\code{paramonte::ParaMonteSampler$readReport()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paramonte" data-topic="ParaMonteSampler" data-id="readRestart">}\href{../../paramonte/html/ParaMonteSampler.html#method-readRestart}{\code{paramonte::ParaMonteSampler$readRestart()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paramonte" data-topic="ParaMonteSampler" data-id="readSample">}\href{../../paramonte/html/ParaMonteSampler.html#method-readSample}{\code{paramonte::ParaMonteSampler$readSample()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiates a \code{ParaDRAM} class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParaDRAM$new(platform = "", website = "")}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a new \code{ParaDRAM} class object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{    pm   = paramonte_class$new()
    pmpd = pm$ParaDRAM()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints \code{ParaDRAM} object usage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParaDRAM$print()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{    pm   = paramonte_class$new()
    pmpd = pm$ParaDRAM()
    pmpd$print()
    # or
}
\if{html}{\out{</div>}}

\if{html}{\out{<div class="r example copy">}}
\preformatted{    pm   = paramonte_class$new()
    pmpd = pm$ParaDRAM()
    pmpd
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParaDRAM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
