% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramonte_class.R
\name{paramonte_class}
\alias{paramonte_class}
\title{Instantiate a \code{\link{R6}} ParaMonte object for \code{\link{paramonteR}} package}
\value{
This function returns a \code{\link{R6}} \code{paramonte_class} object.
}
\description{
\code{paramonte_class$new()} or \code{ paramonte()} initializes a ParaMonte object.\cr
It can then be used to instantiate ParaMonte samplers object.
}
\examples{
pm = paramonte_class$new()

## ------------------------------------------------
## Method `paramonte_class$new`
## ------------------------------------------------

pm = paramonte_class$new()
pm = paramonte()

## ------------------------------------------------
## Method `paramonte_class$print`
## ------------------------------------------------

pm = paramonte()
pm
pm = paramonte()
pm$print()

## ------------------------------------------------
## Method `paramonte_class$ParaDRAM`
## ------------------------------------------------

pm   = paramonte()
pmpd = pm$ParaDRAM()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{paramonte_class$new()}}
\item \href{#method-print}{\code{paramonte_class$print()}}
\item \href{#method-ParaDRAM}{\code{paramonte_class$ParaDRAM()}}
\item \href{#method-clone}{\code{paramonte_class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiates a \code{\link{paramonte_class}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{paramonte_class$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a new \code{\link{paramonte_class}} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm = paramonte_class$new()
}
\if{html}{\out{</div>}}

\if{html}{\out{<div class="r example copy">}}
\preformatted{pm = paramonte()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints \code{\link{paramonte_class}} object usage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{paramonte_class$print()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm = paramonte()
pm
}
\if{html}{\out{</div>}}

\if{html}{\out{<div class="r example copy">}}
\preformatted{pm = paramonte()
pm$print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParaDRAM"></a>}}
\if{latex}{\out{\hypertarget{method-ParaDRAM}{}}}
\subsection{Method \code{ParaDRAM()}}{
This method initializes a ParaDRAM sampler object. It can then be used to call ParaDRAM routines.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{paramonte_class$ParaDRAM()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
This function returns a \code{\link{R6}} \code{ParaDRAM} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm   = paramonte()
pmpd = pm$ParaDRAM()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{paramonte_class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
